Finish local buffer with last changes to database storage
- Change ticks to ulongs in FileDataStorage.cpp
- Initialise ticks with 0 in Mediator.h
- Save participant id as string
- Find intervention mode back properly

+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/FileDataStorage.cpp
@@ -76,7 +76,7 @@ void FileDataStorage::Shutdown() {
 /// @param p_car Current car status in Speed Dreams
 /// @param p_situation Current situation in Speed Dreams
 /// @param p_timestamp Current tick
-void FileDataStorage::Save(tCarElt* p_car, tSituation* p_situation, int p_timestamp) {
+void FileDataStorage::Save(tCarElt* p_car, tSituation* p_situation, unsigned long p_timestamp) {
     WRITE_VAR(m_outputStream, p_timestamp);
     if (m_saveSettings.EnvironmentData) {
         Posd pos = p_car->pub.DynGCg.pos;
